requirements.txt
certifi==2020.6.20
chardet==3.0.4
h5py==2.10.0
idna==2.10
kafka-python==2.0.2
numpy==1.19.2
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
six==1.15.0
urllib3==1.25.10
websocket-client==0.57.0
websockets==8.1
